edgeadm: Move updateControlPlaneInfo form installLiteAPIServer into joinPreparePhase
